requirements.txt
# Git Operations
gitpython>=3.1.40

# Web Application
flask>=3.0.0
flask-cors>=4.0.0

# API Integration (for GitHub data pulling)
requests>=2.31.0

# Data Processing
pandas>=2.0.0
python-dateutil>=2.8.2

# Report Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.3

# LLM Integration - Azure OpenAI
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Async Support
aiohttp>=3.9.0

# Report Generation
jinja2>=3.1.3
markdown>=3.5.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
